+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/custom-blocks/LatexBlock/components/LatexRenderer.tsx
@@ -51,6 +51,7 @@ export const LatexRenderer = ({
         <CodeEditor
           value={formula}
           onChange={onFormulaChange}
+          language="latex"
           onClickOutside={() => setIsLocalEditing(false)}
           parentRef={blockRef}
         />
