+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/custom-blocks/MermaidBlock/types/mermaidTypes.ts
@@ -1,0 +1,12 @@
+export interface MermaidModule {
+  initialize: (config: {
+    startOnLoad: boolean;
+    theme: string;
+    securityLevel: string;
+  }) => void;
+  render: (id: string, text: string) => Promise<{ svg: string; }>;
+}
+
+export interface MermaidImport {
+  default: MermaidModule;
+}
